requirement.txt
python-telegram-bot==21.4
feedparser==6.0.11
APScheduler==3.10.4